imagemanager/requirements.txt
amqp==5.1.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.6.0
attrs==19.3.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
celery==5.2.7
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
contourpy==1.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.8.0
flake8==3.7.9
fonttools==4.38.0
future==0.18.2
httplib2==0.14.0
idna==3.4
importlib-metadata==1.5.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
matplotlib==3.6.2
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
Naked==0.1.31
netifaces==0.10.4
numpy==1.23.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
pyflakes==2.1.1
pygame==2.1.2
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2022.7
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.2
requests-unixsocket==0.2.0
SecretStorage==2.3.1
shellescape==3.8.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.3
systemd-python==234
texttable==1.6.2
twilio==7.16.1
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
websocket-client==0.53.0
xkit==0.0.0
zipp==1.0.0
